Revert "Change format to JSX for better linter"
This reverts commit 157f6642602a0fd353673a37dc1af07e3fe889d1.
